+++ R/netinf.R
@@ -25,7 +25,6 @@
 #' 
 #' @import checkmate
 #' @import assertthat
-#' @import data.table
 #' 
 #' @param  cascades an object of class cascade containing node and cascade 
 #'     information. See \code{\link{as_cascade_long}} and 
@@ -35,12 +34,17 @@
 #' @param params numeric, Parameters for diffusion model. If left unspecified 
 #'     reasonable parameters are inferred from the data. See details for how to 
 #'     specify parameters for the different distributions.
-#' @param n_edges integer or numeric, If integer number of edges to infer per
-#'     iteration, if a numeric value in the interval (0, 1) (excluding 0 and 1) 
-#'     edges are inferred in each iteration until the Vuong test for edge 
-#'     addition reaches the p-value of \code{n_edges} or when the maximum 
-#'     possible number of edges is reached.
+#' @param n_edges integer, number of edges to infer. Leave unspecified if using 
+#'     \code{p_value_cutoff}.
+#' @param p_value_cutoff numeric, in the interval (0, 1). If 
+#'     specifie, edges are inferred in each iteration until the Vuong test for 
+#'     edge addition reaches the p-value cutoff or when the maximum 
+#'     possible number of edges is reached. Leave unspecified if using 
+#'     \code{n_edges} to explicitly specify number of edges to infer.
 #' @param quiet logical, Should output on progress by suppressed.
+#' @param trees logical, Should the inferred cascade trees be returned. Note, 
+#'     that this will lead to a different the structure of the function output. 
+#'     See section Value for details.
 #' 
 #' @return Returns the inferred diffusion network as an edgelist in an object of 
 #'     class \code{diffnet} and \code{\link[base]{data.frame}}. The first 
@@ -55,6 +59,10 @@
 #'             model that have been infered by the approximate profile MLE 
 #'             procedure.
 #'     }
+#'     If the argument \code{trees} is set to \code{TRUE}, the output is a list
+#'     with the first element being the \code{data.frame} described above, and 
+#'     the second element being the trees in edge-list form in a single 
+#'     \code{data.frame}.
 #'  
 #' @references 
 #' M. Gomez-Rodriguez, J. Leskovec, A. Krause. Inferring Networks of Diffusion 
@@ -73,25 +81,29 @@
 #' out <- netinf(cascades2, trans_mod = "exponential", n_edges = 5, params = 1)
 #' 
 #' @export
-netinf <- function(cascades, trans_mod = "exponential", n_edges=0.05, 
-                   params = NULL, quiet = FALSE) {
+netinf <- function(cascades, trans_mod = "exponential", n_edges = NULL, 
+                   p_value_cutoff = NULL, params = NULL, quiet = FALSE, 
+                   trees = FALSE) {
     
     # Check inputs 
     assert_that(class(cascades)[1] == "cascade")
     qassert(trans_mod, "S1")
-    # If no number of edges is specified edge selection is automated via Vuong
-    # Test
-    if(qtest(n_edges, "X1")) {
+    if(is.null(n_edges) & is.null(p_value_cutoff)) {
+        stop('Please specify either `n_edges` or `p_value_cutoff`.')
+    }
+    if(!is.null(n_edges) & !is.null(p_value_cutoff)) {
+        stop('Please only specify either `n_edges` or `p_value_cutoff`.')
+    }
+    if(!is.null(n_edges)) {
         qassert(n_edges, "X1[1,)")
         auto_edges <- FALSE 
         cutoff <- 0 # Not used
-    } else if(qtest(n_edges, "R1")) {
-        qassert(n_edges, "R1(0,1]")
+    } else {
+        qassert(p_value_cutoff, "R1(0,1]")
         auto_edges <- TRUE
-        cutoff <- n_edges
+        cutoff <- p_value_cutoff
         n_edges <- 0 # Not used since n_edges inferred form cutoff
-    } else stop(paste("n_edges has to be either an integer > 1 or a p-value",
-                      "cutoff (0, 1]"))
+    } 
     
     model <- match.arg(trans_mod, c("exponential", "rayleigh", 'log-normal'))
     
@@ -140,10 +152,6 @@ netinf <- function(cascades, trans_mod = "exponential", n_edges=0.05,
     }
     
     # Run netinf 
-    df_cascades <- data.table(as.data.frame(cascades))
-    setkey(df_cascades, "node_name", "cascade_id")
-    network <- data.frame(origin_node = "", destination_node = "")
-    
     netinf_out <- netinf_(cascade_nodes = cascade_nodes, 
                           cascade_times = cascades$cascade_times, 
                           model = model, params = params, n_edges = n_edges, 
@@ -165,7 +173,24 @@ netinf <- function(cascades, trans_mod = "exponential", n_edges=0.05,
     attr(network, "diffusion_model") = model
     attr(network, "diffusion_model_parameters") = params
     
-    return(network) 
+    if(trees) {
+        tree_dfs <- lapply(1:length(netinf_out[[3]]), function(i) {
+            x <- netinf_out[[3]][[i]]
+            out <- as.data.frame(cbind(x[[1]], x[[2]], rep(i, length(x[[1]]))))}
+            )
+        trees_df <- do.call(rbind, tree_dfs)
+        
+        # Replace int node ids with node_names 
+        trees_df$child <- do.call(c, cascades$cascade_nodes)
+        trees_df <- trees_df[!is.na(trees_df[, 2]), ]
+        trees_df[, 1] <- cascades$node_names[(trees_df[, 1] + 1)]
+        casc_names <- names(cascades$cascade_nodes)
+        trees_df[, 3] <- casc_names[trees_df[, 3]] 
+        colnames(trees_df) <- c("parent", "log_score", "cascade_id", "child") 
+        trees_df <- trees_df[, c(1, 4, 2, 3)]
+        return(list('network' = network, 'trees' = trees_df))
+    }
+    else return(network) 
 }